Clearing Stack
Added a functionality of clearing the stack after "Okay :/" button is
pressed on the credits screen, using Intent.FLAG_ACTIVITY_NEW_TASK and
Intent.FLAG_ACTIVITY_CLEAR_TASK. Now, tapping the "Okay :/" button leads
back to the home screen and tapping the Android back button leads to the
Android Home Screen, instead of the app's splash screen. This
functionality should be added to the application home screen as well,
when it is opened directly.