Successful Exits
Added proper exit support for both Android Back Button and also the
"Exit" button in the main menu of the app. However, the "Exit" button in
the app's main menu does not prompt if the user really wants to exit.
Fixed the "Okay :/" button the app's main menu. Clicking it will simply
end the "credits" activity. It took me this long to figure it that it
can be done using finish().