Disabled Rotation
Finally disabled screen rotation when the app is running. Screen
rotation resets the ball's state in the app and jeopardizes the main
menu's layout. Again, it took me this long to realize that
android:screenOrientation="potrait" attribute in the AndroidManifest.xml
will do the job just right. The main game screen isn't an Activity. It's
a Canvas, which needs another Activity class with an XML layout to run.
Hence, android:screenOrientation="potrait" attribute is added to this
auxiliary screen layout that I created.