First Phase of Success
Achieved the "teleportation" effect. The ball comes to the point on the
screen that the user touches but does it instantaneously. I tried hard
to increase the time taken by the ball to reach the destination but
couldn't succeed and accidentally invented this teleportation. Whenever
a user taps on the screen, the ball moves to that point from its current
location at a very high speed, which looks like teleportation. Also
fixed a couple of bugs that arose due the float nature of
Canvas.getHeight() and Canvas.getWidth() methods. I previously fixed
horizontal movement by adding an integer to decrease the limit of
comparison but the value of that integer was too high so I reduced it.
Also reduced the integer to be subtracted from Canvas.getWidth() method
to avoid the only-vertical-motion anomaly. Added four new variables,
'a', 'b', 'factX' and 'factY'. 'a' and 'b' are for storing the
coordinates of the point where the screen is touched by the user.
'factX' and 'factY' are the factors on X and Y axes respectively, by
which the values of 'x' and 'y' should respectively be incremented or
decremented corresponding to the location of the touched point with
respect to the location of the ball's centre, in order for the ball to
reach the touched location.