Disco Ball
Temporarily disabled the "teleportation" and reverted the app's function
to as it was before. However, added a feature that I thought of,
initially, that a cousin of mine pointed out, might be good. The ball
starts moving with the first tap on the screen, with an initial colour
of blue(hexadecimal code #0099cc) and will change colour each time it
hits a wall, or, the edge of the screen.  Each edge is assigned a colour
in hexadecimal as you can see in the code.